+++ www/types/index.d.ts
@@ -1,3 +1,4 @@
 export { Configuration } from "./Configurations";
 export { CreateOptions, CreateResult } from "./CreateConference";
 export { JoinOptions, JoinUserInfo, UserType } from "./JoinConference";
+export { ConferenceStatus, ConferenceStatusUpdated } from "./ConferenceStatus";